+++ tests/subsys/sd/sdmmc/prj.conf
@@ -2,5 +2,6 @@ CONFIG_TEST=y
 CONFIG_ZTEST=y
 CONFIG_SDHC=y
 CONFIG_SD_STACK=y
+CONFIG_SDMMC_STACK=y
 CONFIG_LOG=y
 CONFIG_ZTEST_NEW_API=y
